+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/bot/BotCommand.kt
@@ -1,6 +1,6 @@
 package dev.inmo.tgbotapi.requests.bot
 
-import dev.inmo.tgbotapi.CommonAbstracts.WithOptionalLanguageCode
+import dev.inmo.tgbotapi.types.abstracts.WithOptionalLanguageCode
 import dev.inmo.tgbotapi.requests.abstracts.SimpleRequest
 import dev.inmo.tgbotapi.types.commands.BotCommandScope