+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/abstracts/WithLanguageCode.kt
@@ -1,4 +1,4 @@
-package dev.inmo.tgbotapi.CommonAbstracts
+package dev.inmo.tgbotapi.types.abstracts
 
 import dev.inmo.micro_utils.language_codes.IetfLanguageCode