+++ tgbotapi.core/src/jvmMain/kotlin/dev/inmo/tgbotapi/types/UserLocale.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.types
 
 import dev.inmo.micro_utils.language_codes.IetfLanguageCode
-import dev.inmo.tgbotapi.CommonAbstracts.WithLanguageCode
+import dev.inmo.tgbotapi.types.abstracts.WithLanguageCode
 import java.util.*
 
 fun IetfLanguageCode?.javaLocale() = this ?.code ?.let { Locale.forLanguageTag(it) }